requirements.txt
python-dateutil
feedparser
requests
PyYAML
django
psycopg2-binary
